feat: Implement Q-learning driving agent in agent.py
epsilon_decay and gamma are added as class members to keep track of the
epsilon decay rate and the gamma factor which is fixed at 1 for now.

epsilon is decayed linearly before each trial

A private method has been created to help compute the best action and Q-value
for any particular state since this is needed in a few places.

get_maxQ() is implemented to return the best Q-value for a state

createQ() is implemented to initialize the Q-values for each state-action pair

choose_action() is implemented to pick either a random action based on epsilon
or choose the action with the highest Q-value

learn() is implemented to update the Q-values based on the Q-learning update
rule